Improve performance of the pixel-perfect scaling in Magnifier
It seems that moving mouse a lot can sometimes stop the magnifier
from moving until the cursor slows down again.
This seems to be (mostly?) limited to running the app under debugger
and is hard to replicate/notice in normal circumstances.
Based on the debugger's graphs, it seems to be triggered by the runtime
making A LOT of garbage collection calls.